# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.UserInfoResponse:
    properties:
      balance:
        example: 100.5
        type: number
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      tariff_id:
        type: integer
      type:
        $ref: '#/definitions/models.TransactionType'
      user_id:
        type: integer
    type: object
  models.TransactionType:
    enum:
    - add
    - buy
    type: string
    x-enum-varnames:
    - Add
    - Buy
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample API with Swagger
  title: API Computer-club
  version: "1.0"
paths:
...
  /users:
    get:
      description: Получение списка пользователей (может только админ)
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей успешно показан
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            type: string
        "404":
          description: Список пользователей пуст
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение списка пользователей
      tags:
      - users
  /users/{id}:
    get:
      description: Получение пользователя по ID (может только админ)
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно показан
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректный id
          schema:
            type: string
        "403":
          description: Недостаточно прав
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение пользователя по ID
      tags:
      - users
  /users/info:
    get:
      description: Возвращает информацию о пользователе, баланс и транзакции
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserInfoResponse'
        "400":
          description: Некорректный user_id
          schema:
            type: string
        "401":
          description: Кошелек не найден или пользователь
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение информации о пользователе
      tags:
      - users
swagger: "2.0"